Distribute code from state_menu across 4 corresponding files
