+++ SAUCE/gamestates/state_demo.h
@@ -244,7 +244,7 @@ const unsigned char Credits2[629]={
 	#ifdef FLAG_BETA_BUILD
 		const unsigned char ver[] = "DEMO";
 	#else
-		const unsigned char ver[] = "DEMO";
+		const unsigned char ver[] = " VER";
 	#endif
 #endif
