+++ SAUCE/include_menus.h
@@ -1,0 +1,14 @@
+
+CODE_BANK_PUSH("XCD_BANK_03")
+
+#include "defines/menunametable.h"
+#include "menustates/levelselection.c"
+#include "menustates/progressbar.c"
+#include "menustates/customize_screen.c"
+#include "menustates/titlescreen.c"
+
+CODE_BANK("XCD_BANK_05")
+CODE_BANK("XCD_BANK_06")
+CODE_BANK("XCD_BANK_02")
+
+CODE_BANK_POP()
