+++ SAUCE/menustates/titlescreen.c
@@ -5,9 +5,9 @@ CODE_BANK_PUSH("XCD_BANK_03")
 void color_picker();
 #endif
 void check_if_music_stopped();
-void clear_shit();
 void movement();
 void bounds_check();
+
 void title_cube_shit();
 void title_robot_shit();
 void title_mini_wave_shit();
@@ -16,8 +16,7 @@ void title_ufo_shit();
 void title_ball_shit();
 void title_swing_shit();
 void title_ship_shit();
-void refreshmenu();
-void refreshmenu_part2();
+
 void loop_routine_update();
 void dec_mouse_timer();
 void roll_new_mode();
@@ -27,10 +26,7 @@ void state_demo();
 void mouse_and_cursor();
 void colorinc();
 void colordec();
-void leveldec();
-void levelinc();
 void set_settings();
-void start_the_level();
 extern uint8_t famistudio_song_speed;
 
 const uint8_t xbgmlookuptable[];
@@ -117,619 +113,6 @@ const uint8_t G_Table2[]={
 	0x2B,
 	0x2C
 };
-
-#include "charmaps/mainmenu_charmap.h"
-
-const uint8_t lvlselect_irq_table[] = {
-	31,
-	irqtable_hscroll, 0x00, // item 0x02 gets modified every frame
-	149,
-	irqtable_hscroll, 0x00,
-
-	irqtable_end // always end with 0xff
-};
-void draw_both_progress_bars();
-void levelselection() {
-	if (tempplat == 1) { tempplat = 0; options |= platformer; }
-	mmc3_set_8kb_chr(MENUBANK);
-	mmc3_set_2kb_chr_bank_0(0xFF);
-	mmc3_set_2kb_chr_bank_1(MOUSEBANK);
-	pal_fade_to_withmusic(4,0);
-	mmc3_disable_irq();
-	disco_sprites = 0;
-
-	pal_bg(oldsplashMenu);
-	write_irq_table(lvlselect_irq_table);
-	set_irq_ptr(irqTable);
-	oam_clear();
-	ppu_off();
-	pal_bright(0);
-	// pal_bg(paletteMenu);
-	set_scroll_x(0);
-	set_scroll_y(0);  
-
-	#if __VS_SYSTEM
-		menutimer = 0;
-		if (!coins_inserted) {
-			pal_bg(gameoverpalette);
-			vram_adr(NAMETABLE_A);
-			vram_unrle(gameover); 
-			vram_adr(NAMETABLE_B);
-			vram_unrle(gameover);
-			tmp5 = 200;
-			ppu_on_all();		
-	//		kandowatchesyousleep = 0;
-	//		exitingLevelSelect = 1;
-	//		cube_data[0] = 0;
-	//		cube_data[1] = 0;
-	//		return;
-		} else {
-	#endif
-
-#if !__VS_SYSTEM
-	vram_adr(NAMETABLE_A);
-	vram_unrle(level_select_screen); 
-	vram_adr(NAMETABLE_B);
-	vram_unrle(level_select_screen);
-#else
-	vram_adr(NAMETABLE_A);
-	vram_unrle(game_main_menu); 
-	vram_adr(NAMETABLE_B);
-	vram_unrle(game_main_menu);
-#endif
-	memfill(attemptCounter, 0, sizeof(attemptCounter));
-
-	tmp8 = 0xff00;
-	edit_irq_table(high_byte(tmp8),2);
-	tmp4 = 1;
-	crossPRGBankJump0(refreshmenu);
-	crossPRGBankJump0(refreshmenu_part2);
-	top_triggers = 0;
-	triggers = 0;
-	triggers_hit[0] = 0;
-	triggers_hit[1] = 0;
-	triggers_hit[2] = 0;
-
-	// one_vram_buffer(0xb0+kandokidshackTENS, NTADR_A(17,17));
-	// one_vram_buffer(0xb0+kandokidshackONES, NTADR_A(18,17));
-
-	cube_rotate[0] = 0;
-	cube_rotate[1] = 0;
-
-	#if __VS_SYSTEM
-	if (menuMusicCurrentlyPlaying == 0 && !nestopia) music_play(idx8_load(xbgmlookuptable, newrand() & 31));
-	#else
-	if (menuMusicCurrentlyPlaying == 0 && !nestopia) music_play(xbgmlookuptable[menu_music]);
-	#endif
-	menuMusicCurrentlyPlaying = 1;
-
-	ppu_on_all();
-	drawBarFlag = 2;
-	low_byte(tmpA) = 1;
-	draw_both_progress_bars();
-	
-	#if __VS_SYSTEM
-	}	// end the else statement
-	#endif
-
-	ppu_wait_nmi();
-	ppu_wait_nmi();
-	pal_fade_to_withmusic(0,4);
-
-	#if __VS_SYSTEM
-		if (!coins_inserted) {
-			while (tmp5) {
-				ppu_wait_nmi();
-				music_update();
-				tmp5--;
-			}
-			ppu_off();
-			kandowatchesyousleep = 0;
-			exitingLevelSelect = 1;
-			cube_data[0] = 0;
-			cube_data[1] = 0;
-			return;
-		} else {
-	#endif
-
-	while (1){
-		loop_routine_update();
-		 // read the first controller
-		check_if_music_stopped();
-		// scroll
-		if (tmp4) edit_irq_table(high_byte(tmp8),2);
-		else edit_irq_table(low_byte(tmp8),2);
-		
-		if (low_byte(tmp8) > 0) {
-			low_byte(tmp8) -= (low_byte(tmp8) >> 2) + 1;
-		}
-		
-		high_byte(tmp8) = low_byte(tmp8)^0xff;
-		
-		if (mouse.left_press) {
-			if (mouse.y >= 0x6E && mouse.y <= 0x7B) {
-				if (mouse.x >= 0x09 && mouse.x <= 0x15) {
-					drawBarFlag = 2;
-					leveldec();
-				}
-				else if (mouse.x >= 0xE5 && mouse.x <= 0xEF) {
-					drawBarFlag = 2;
-					levelinc();				
-				}
-			}
-			if (((mouse.y >= 0x3D && mouse.y <= 0x6C) && (mouse.x >= 0x2D && mouse.x <= 0xCC))) {
-				start_the_level();
-				return;				
-			}
-			
-		}
-		// no longer required because of the menu option 
-		//if (twoplayer) one_vram_buffer('d', NTADR_A(31, 2));
-		//else one_vram_buffer('e', NTADR_A(31, 2));
-
-		//if (pad[0] & PAD_UP && pad_new[0] & PAD_SELECT) { twoplayer ^= 0x01; sfx_play(sfx_coin, 0); }
-
-		#if __VS_SYSTEM
-			menutimer++;
-			if (menutimer >= 2000) {
-				exitingLevelSelect = 1;
-				kandowatchesyousleep = 0;
-				return;
-			}
-		#endif
-
-		if (joypad1.press & (PAD_START | PAD_A)){
-			start_the_level();
-			#if __VS_SYSTEM
-				menutimer = 0;
-			#endif
-			return;
-		}
-
-		if (joypad1.press_b){
-			exitingLevelSelect = 1;
-			kandowatchesyousleep = 0;
-			#if __VS_SYSTEM
-				menutimer = 0;
-			#endif
-			return;
-		}
-		
-		if (joypad1.press_right){
-			drawBarFlag = 2;
-			levelinc();
-			#if __VS_SYSTEM
-				menutimer = 0;
-			#endif
-		}
-		if (joypad1.press_left){
-			drawBarFlag = 2;
-			leveldec();
-			#if __VS_SYSTEM
-				menutimer = 0;
-			#endif		
-		}
-
-		// NOTE: this is hardcoded. It used to be 0x3F with the old scroll scheme
-		if (tmp8 == 0x3B) {
-			crossPRGBankJump0(refreshmenu);
-			crossPRGBankJump0(refreshmenu_part2);
-		}
-
-		draw_both_progress_bars();
-
-		dec_mouse_timer();
-	}
-	#if __VS_SYSTEM	
-	}	// end the else statement
-	#endif
-
-}
-
-
-void draw_progress_bar();
-void draw_full_progress_bar();
-void draw_percentage();
-
-/*
-	all 8 bit tmps:
-
-	tmp1 : x
-	tmp2 : y
-	tmp3 : offset into sprite table
-	tmp4 : needed inside the transitions
-	tmp7 : percentage
-	
-	free 8 bit tmps inside calls:
-	- tmp8
-	damn...
-
-	technically i can use the low byte of 16 bits tmps
-*/
-void draw_both_progress_bars() {
-	
-		// Normal level completeness stuff
-		tmp7 = level_completeness_normal[level];
-		hexToDec(tmp7);
-		tmp1 = 112;		// x
-		if (tmp4) {
-			tmp1 += tmp8 + 1;
-		} else {
-			tmp1 -= tmp8;
-		}
-		
-		tmp2 = 127;		// y
-		tmp3 = 0;		// offset
-		draw_percentage();	
-
-		tmp1 = 6; 		// x
-		tmp2 = 16;		// y
-		low_byte(tmp6) = 1; 		// index
-		draw_full_progress_bar();
-		
-	
-		// Practice level completeness stuff
-		tmp7 = level_completeness_practice[level];
-		hexToDec(tmp7);
-		tmp1 = 112;		// x
-		if (tmp4) {
-			tmp1 += tmp8 + 1;
-		} else {
-			tmp1 -= tmp8;
-		}
-		tmp2 = 151;		// y
-		tmp3 = 0;		// offset
-		draw_percentage();	
-
-		tmp1 = 6; 		// x
-		tmp2 = 19;		// y
-		low_byte(tmp6) = 2; 		// index
-		draw_full_progress_bar();
-
-		
-}
-void calculate_sprite_pos();
-void draw_bar_sprites() {
-	#define temp_save animating
-	// Put percentage mask
-	tmp3 = 0;
-	tmp1--; // left 1 tile
-
-	// convert tile cords into sprite cords (basically multiply by 8 aka shift left 3 times)
-	tmp1 = (tmp1 << 3) + 1;
-	temp_save = tmp1;
-	tmp2 = (tmp2 << 3) - 1;
-	if (tmp4) {
-		tmp1 += tmp8 + 1;
-	} else {
-		tmp1 -= tmp8;
-	}
-	crossPRGBankJump0(put_progress_bar_sprite);
-	tmp1++;
-	tmp3 = 3;
-	crossPRGBankJump0(put_progress_bar_sprite);
-
-	tmp1 = 206;
-	if (tmp4) {
-		tmp1 += tmp8 + 1;
-	} else {
-		tmp1 -= tmp8;
-	}
-	tmp3 = 4;
-	crossPRGBankJump0(put_progress_bar_sprite);
-	tmp1 = temp_save;
-	crossPRGBankJump0(calculate_sprite_pos);
-	
-	if (tmp4) {
-		tmp1 += tmp8 + 1;
-	} else {
-		tmp1 -= tmp8;
-	}
-	
-	tmp3 = low_byte(tmp6);
-	crossPRGBankJump0(put_progress_bar_sprite);
-	temp_save = 0;
-}
-
-void draw_full_progress_bar() {
-	if (drawBarFlag) crossPRGBankJump0(draw_progress_bar);
-	// end tmp3 need
-
-	// tmp5 is now used for the temporary sprite x pos with subpixels
-	draw_bar_sprites();
-	// end tmp5 usage
-}
-
-void draw_level_progress() {
-	update_level_completeness();
-	hexToDec(__A__);
-	tmp1 = 116;		// x
-	tmp2 = 0;		// y
-	tmp3 = 25;		// offset
-	draw_percentage();	
-}
-
-void draw_percentage() {
-	if (hexToDecOutputBuffer[2]) { 				  
-		high_byte(tmp6) = hexToDecOutputBuffer[2]; // number
-		crossPRGBankJump0(put_number);
-	}
-
-	if (hexToDecOutputBuffer[2] | hexToDecOutputBuffer[1]) {
-		tmp1 += 8;					    		   // x
-		high_byte(tmp6) = hexToDecOutputBuffer[1]; // number
-		crossPRGBankJump0(put_number);
-	}
-
-	tmp1 += 8;					   			   // x
-	high_byte(tmp6) = hexToDecOutputBuffer[0]; // number
-	crossPRGBankJump0(put_number);
-
-	tmp1 += 8;					    // x
-	high_byte(tmp6) = 10; 			// percentage
-	crossPRGBankJump0(put_number);
-}
-
-#include "charmaps/mainmenu_customize.h"
-
-const uint8_t loNTAddrTableCustomizeScreen[] = {
-	LSB(NTADR_A(13, 8)),		// -1 = 3
-	LSB(NTADR_A(4, 14)),		// 0
-	LSB(NTADR_A(13, 14)),		// 1
-	LSB(NTADR_A(22, 14)),		// 2
-	LSB(NTADR_A(13, 8)),		// 3
-	LSB(NTADR_A(4, 14))			// 4 = 0
-};
-
-const uint8_t hiNTAddrTableCustomizeScreen[] = {
-	MSB(NTADR_A(13, 8)),		// -1 = 3
-	MSB(NTADR_A(4, 14)),		// 0
-	MSB(NTADR_A(13, 14)),		// 1
-	MSB(NTADR_A(22, 14)),		// 2
-	MSB(NTADR_A(13, 8)),		// 3
-	MSB(NTADR_A(4, 14))			// 4 = 0
-};
-
-void updateColors() {
-	if (joypad1.press_up) {
-		crossPRGBankJump0(colorinc);
-	}
-	if (joypad1.press_down) { 
-		crossPRGBankJump0(colordec);
-	}
-	//if ((pad[0] & PAD_SELECT) && (pad_new[0] & PAD_A)) icon_colors[settingvalue] = 0x0D;	
-}
-
-void customize_screen() {
-#define prev_icon tmp8
-	settingvalue = 3; 
-	pal_fade_to_withmusic(4,0);
-	mmc3_disable_irq();
-	ppu_off();
-	pal_bg(paletteMenu);
-	pal_col(0x00, 0x00);
-	pal_set_update();
-	mmc3_set_8kb_chr(MENUICONBANK);
-	mmc3_set_2kb_chr_bank_0(0xFF);
-	mmc3_set_2kb_chr_bank_1(MOUSEBANK);
-	vram_adr(NAMETABLE_A);
-	vram_unrle(customizescreen);   	
-	#if __VS_SYSTEM
-	menutimer = 0;
-	#endif
-	kandokidshack = 0;
-	kandokidshack2 = 0;
-	tmp2 = 0;
-	do {
-		// kandokidshack = kandokidshack + coin1_obtained[tmp2] + coin2_obtained[tmp2] + coin3_obtained[tmp2];
-		__A__ = tmp2; __asm__("tay");
-		__A__ = kandokidshack;
-		__asm__("clc \n adc %v, y", coin1_obtained);
-		__asm__("clc \n adc %v, y", coin2_obtained);
-		__asm__("clc \n adc %v, y", coin3_obtained);
-		kandokidshack = __A__;
-		if (LEVELCOMPLETE[tmp2]) kandokidshack2 += stars_list[tmp2];
-
-		tmp2++;
-	}
-	#ifdef FLAG_ENABLE_TEST_LEVELS
-	while (tmp2 < 255);
-	#else
-	while (tmp2 < LEVEL_COUNT2); 
-	#endif
-
-	
-	printDecimal(kandokidshack2, 2, 0xD0, 0xFF, NTADR_A(26,4));
-	printDecimal(kandokidshack, 2, 0xD0, 0xFF, NTADR_A(26,5));
-
-	one_vram_buffer('h', NTADR_A(13, 8));		
-	one_vram_buffer('i', NTADR_A(13, 9));
-	one_vram_buffer('f', NTADR_A(18, 8));
-	one_vram_buffer('g', NTADR_A(18, 9));
-	prev_icon = !icon;
-	ppu_on_all();
-	pal_fade_to_withmusic(0,4);
-	tmp1 = iconTable[icon] + 'a';
-	one_vram_buffer(tmp1, NTADR_A(15, 8));
-	one_vram_buffer(++tmp1, NTADR_A(16, 8));
-	one_vram_buffer((tmp1 += ('c'-'b')), NTADR_A(15, 9));
-	one_vram_buffer(++tmp1, NTADR_A(16, 9));
-	prev_icon = icon;	
-	while (1) {
-		tmp3 = 0;
-		check_if_music_stopped();
-		loop_routine_update();
-
-		pal_col(0x0a,color1);
-		pal_col(0x0b,color2);
-		pal_col(0x09,color3);
-		pal_set_update();
-		 // read the first controller
-	
-//mouse stuff
-
-	#if !__VS_SYSTEM
-		if (mouse.left_press || (mouse.left && hold_timer >= 15)) {
-//icon
-			if ((mouse.x >= 0x76 && mouse.x <= 0x83)) { 
-				if (mouse.y >= 0x34 && mouse.y <= 0x3C) {
-					icon++;
-					if (icon > (MAX_ICONS - 1)) icon = 0;				
-					settingvalue = 3;
-					hold_timer = 0;					
-					clear_shit();
-				}
-				else if (mouse.y >= 0x4D && mouse.y <= 0x54) {
-					if (icon == 0) icon = MAX_ICONS - 1;
-					else icon--;				
-					settingvalue = 3;
-					hold_timer = 0;					
-					clear_shit();
-					
-				}
-//color 2
-				else if ((mouse.y >= 0x64 && mouse.y <= 0x6C)) {
-					settingvalue = 1;
-					crossPRGBankJump0(colorinc);
-					hold_timer = 0;					
-					clear_shit();
-					
-				}
-				else if ((mouse.y >= 0x7D && mouse.y <= 0x83)) {
-					settingvalue = 1;			
-					crossPRGBankJump0(colordec);
-					hold_timer = 0;					
-					clear_shit();
-					
-				}
-
-				
-			}
-//color 1
-			else if ((mouse.x >= 0x2E && mouse.x <= 0x3B)) {
-				if (mouse.y >= 0x63 && mouse.y <= 0x6C) {
-					settingvalue = 0;
-					crossPRGBankJump0(colorinc);
-					hold_timer = 0;					
-					clear_shit();
-					
-				}
-				else if ((mouse.y >= 0x7D && mouse.y <= 0x83)) {
-					settingvalue = 0;			
-					crossPRGBankJump0(colordec);
-					hold_timer = 0;					
-					clear_shit();
-					
-				}
-			}
-//color3
-			else if ((mouse.x >= 0xBE && mouse.x <= 0xCB)) {
-				if (mouse.y >= 0x63 && mouse.y <= 0x6C) {
-					settingvalue = 2;
-					crossPRGBankJump0(colorinc);
-					hold_timer = 0;					
-					clear_shit();
-					
-				}
-				else if ((mouse.y >= 0x7D && mouse.y <= 0x83)) {
-					settingvalue = 2;			
-					crossPRGBankJump0(colordec);
-					hold_timer = 0;					
-					clear_shit();
-					
-				}
-			}
-			if ((mouse.x >= 0x35 && mouse.x <= 0xBC) && (mouse.y >= 0xBC && mouse.y <= 0xC4)) {		
-				return;			//go back
-			}
-		}
-	#endif
-//end mouse stuff
-		if (!retro_mode) {
-			if (icon != prev_icon) {
-				tmp1 = iconTable[icon] + 'a';
-				one_vram_buffer(tmp1, NTADR_A(15, 8));
-				one_vram_buffer(++tmp1, NTADR_A(16, 8));
-				one_vram_buffer((tmp1 += ('c'-'b')), NTADR_A(15, 9));
-				one_vram_buffer(++tmp1, NTADR_A(16, 9));
-				prev_icon = icon;
-			}
-		} else {
-			if (prev_icon) {
-				one_vram_buffer('a'+0x6E, NTADR_A(15,8));
-				one_vram_buffer('b'+0x6E, NTADR_A(16,8));
-				one_vram_buffer('c'+0x6E, NTADR_A(15,9));
-				one_vram_buffer('d'+0x6E, NTADR_A(16,9));
-				prev_icon = 0;
-			}
-		}
-		if (settingvalue == 3 && !retro_mode) {
-			if (joypad1.press_up || (joypad1.up && hold_timer >= 15)) {
-				icon++;
-				#if __VS_SYSTEM
-				menutimer = 0;
-				#endif
-				if (icon > (MAX_ICONS - 1)) icon = 0;
-				hold_timer = 0;
-			}
-			if (joypad1.press_down || (joypad1.down && hold_timer >= 15)) {
-				#if __VS_SYSTEM
-				menutimer = 0;
-				#endif
-				if (icon == 0) icon = MAX_ICONS - 1;
-				else icon--;
-				hold_timer = 0;				
-			}
-		} else if (settingvalue != 3) updateColors();
-
-		if (joypad1.press_right) {
-			settingvalue++;
-			#if __VS_SYSTEM
-			menutimer = 0;
-			#endif
-			if (settingvalue == 4) settingvalue = 0;
-			tmp3--;
-		}
-		if (joypad1.press_left) {
-			if (settingvalue == 0) settingvalue = 3;
-			else settingvalue--;
-			#if __VS_SYSTEM
-			menutimer = 0;
-			#endif
-			tmp3++;
-		}
-
-		if (tmp3) {
-			tmp4 = settingvalue; ++tmp4;
-			tmp5 = loNTAddrTableCustomizeScreen[tmp4]|(hiNTAddrTableCustomizeScreen[tmp4]<<8);
-			one_vram_buffer('h', tmp5);		
-			one_vram_buffer('i', tmp5 + VRAM_OFF(0, 1));
-			one_vram_buffer('f', tmp5 = addloNOC(tmp5, VRAM_OFF(5, 0)));
-			one_vram_buffer('g', tmp5 + VRAM_OFF(0, 1));
-
-			tmp4 += tmp3;   // Get the old index
-			tmp5 = loNTAddrTableCustomizeScreen[tmp4]|(hiNTAddrTableCustomizeScreen[tmp4]<<8);
-			one_vram_buffer(' ', tmp5);
-			one_vram_buffer(' ', tmp5 + VRAM_OFF(0, 1));
-			one_vram_buffer(' ', tmp5 = addloNOC(tmp5, VRAM_OFF(5, 0)));
-			one_vram_buffer(' ', tmp5 + VRAM_OFF(0, 1));
-		}
-
-		if (joypad1.press_b) {
-			return;
-		}
-		dec_mouse_timer();
-		hold_timer++;
-		#if __VS_SYSTEM
-		menutimer++;
-		if (menutimer >= 2000) return;						
-		#endif
-	}
-#undef prev_icon
-}
-
-
-
 
 const uint8_t UFO_Title_Jump_Table[]={
 	5,
@@ -1741,61 +1124,6 @@ void state_menu() {
 	
 }
 
-
-void leveldec() {
-	--level;
-	//if (level == 0x0B) level = 0x0A;	//THEORY OF EVERYTHING SKIP
-	low_byte(tmp8) = 0xff;
-	tmp4 = 0;
-	if (!normalorcommlevels) {
-		if (level == 0xFF){
-			level = LEVEL_COUNT-1;
-			
-		}
-	}
-	else {
-		if (level < LEVEL_COUNT) level = LEVEL_COUNT2 - 1;
-	}
-	//break;
-	crossPRGBankJump0(refreshmenu);
-}			
-
-void levelinc() {
-	++level;
-	//if (level == 0x0B) level = 0x0C;	//THEORY OF EVERYTHING SKIP
-	low_byte(tmp8) = 0xff;
-	tmp4 = 1;
-	if (!normalorcommlevels) {
-		if (level >= LEVEL_COUNT){
-			level = 0x00;
-			
-		}
-	}
-	else {
-		if (level >= LEVEL_COUNT2){
-			level = LEVEL_COUNT;
-		}
-	}
-	crossPRGBankJump0(refreshmenu);
-}			
-
-void start_the_level() {
-	sfx_play(sfx_start_level, 0);
-	famistudio_music_stop();
-	if (tmp4) edit_irq_table(high_byte(tmp8),2);
-	else edit_irq_table(low_byte(tmp8),2);
-	tmpA = 0;
-	do {
-		oam_clear();
-		draw_both_progress_bars();
-		ppu_wait_nmi();
-		music_update();
-	} while (++tmpA < 30);
-	gameState = 0x02;
-	pal_fade_to(4,0);
-	menuMusicCurrentlyPlaying = 0;
-}			
-
 void set_title_icon() {
 		if (titlemode < 4) {
 			while (tmp7 > 26) {
@@ -2018,25 +1346,6 @@ void title_robot_shit() {
 		if (currplayer_y_small >= 160) { currplayer_gravity = GRAVITY_DOWN; tmpi8 = 0; teleport_output = 0; currplayer_y_small = 160; }
 	}
 					
-}
-void clear_shit() {
-	one_vram_buffer(' ', NTADR_A(4,14));	
-	one_vram_buffer(' ', NTADR_A(4,15));	
-	one_vram_buffer(' ', NTADR_A(9,14));	
-	one_vram_buffer(' ', NTADR_A(9,15));	
-	one_vram_buffer(' ', NTADR_A(13,14));	
-	one_vram_buffer(' ', NTADR_A(13,15));	
-	one_vram_buffer(' ', NTADR_A(18,14));	
-	one_vram_buffer(' ', NTADR_A(18,15));	
-	one_vram_buffer(' ', NTADR_A(22,14));	
-	one_vram_buffer(' ', NTADR_A(22,15));	
-	one_vram_buffer(' ', NTADR_A(27,14));	
-	one_vram_buffer(' ', NTADR_A(27,15));	
-	one_vram_buffer(' ', NTADR_A(13,8));	
-	one_vram_buffer(' ', NTADR_A(13,9));	
-	one_vram_buffer(' ', NTADR_A(18,8));	
-	one_vram_buffer(' ', NTADR_A(18,9));	
-	
 }
 
 void check_if_music_stopped() {
